+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Release TBD
   * Fix setting enabled flag for plugin products.
   * Return to editing page when adding a product and the image uploads fail.
   * Fix showing previous item's image on order when next item has no image.
+  * Clear sitemap cache when changing products or categories.
 
 ## v0.7.0
 Release 2019-06-02